Move SSO flow into connect spec; fix WPDashboard issue
