+++ lib/pages/wp-admin/wp-admin-dashboard-page.js
@@ -14,7 +14,7 @@ export default class WPAdminDashboardPage extends BaseContainer {
 			visit = true;
 		}
 
-		super( driver, By.css( '#wpbody .welcome-panel-content' ), visit, wpAdminURL );
+		super( driver, By.css( '#wpbody #wpbody-content' ), visit, wpAdminURL );
 	}
 
 	logout() {
